@@ -48,7 +48,7 @@ setup(
         "ml": ["emmet-core[ml]"],
         "openmm": ["transport-analysis>=0.1.0"],
     },
-    python_requires=">=3.9",
+    python_requires=">=3.10",
     license="modified BSD",
     zip_safe=False,
 )
