@@ -1,4 +1,4 @@
-from typing import Literal, List, Union, Optional
+from typing import List, Literal, Optional, Union
 
 from pydantic import Field
 from pymatgen.analysis.chemenv.coordination_environments.chemenv_strategies import (
@@ -14,8 +14,7 @@ from pymatgen.analysis.chemenv.coordination_environments.structure_environments 
     LightStructureEnvironments,
 )
 from pymatgen.analysis.structure_analyzer import SpacegroupAnalyzer
-from pymatgen.core.structure import Molecule
-from pymatgen.core.structure import Structure
+from pymatgen.core.structure import Molecule, Structure
 
 from emmet.core.material_property import PropertyDoc
 from emmet.core.mpid import MPID
@@ -377,7 +376,7 @@ class ChemEnvDoc(PropertyDoc):
     def from_structure(
         cls,
         structure: Structure,
-        material_id: MPID,
+        material_id: MPID | None = None,
         **kwargs,
     ):  # type: ignore[override]
         """
